Implementation:  Moved responsibility for saving tags out of PaperHandler. Implemented FileTagHandler for saving new file tags from imported files in the database.
